Introduce utils.py and fix check for positive quantities
Now we can collect utility functions that are not directly related
to calculations in utils.py.

For example, check_pos_params is now put there, as well as
pint_append.
